requirements.txt
pytesseract
Pillow
pyinstaller
openai
python-telegram-bot
pyupdater
matplotlib
